Allow `Lotus::View::Template` to accept one or two arguments
The first one is the template path, while the second is the optional
encoding.

For production code, the encoding is always specified. The value to look
at is `Lotus::View.configuration.default_encoding`.

I made it optional only for testing purposes.
When we generate a view test in `lotusrb`, it prepares the template like
this:

  # ...
  describe Web::Views::Home::Index do
    # ...
    let(:template)  { Lotus::View::Template.new('apps/web/templates/home/index.html.erb') }
  end

It's annoying to be forced to specify the encoding in this case. Most of
the applications out there just use UTF-8. So we are simplifying the
most common case.

/cc @lotus/core-team