Render template content before rendering layout content
This allows for content sharing from the template up to the layout. For example, via a “content_for”-style method, which captures content provided by the template and allows the layout to access it (useful for e.g. specifying page-specific content that needs to go into the layout’s `<head>`).