Add Scope to act as root template scope and clarify partial rendering behavior
Make the part available to a partial by its own name, rather than the part _becoming_ the scope for the partial.

This allows the partial to be authored in a more conventional manner. So for a partial rendered on a part like `user.render :info`, you can now reference the user in the partial by name:

= user.email

Instead of direct, implicit access to the part’s methods:

= email

This also makes things clearer when additional locals are provided when rendering on a part, e.g. `user.render :info, title :”Info box”`. Accessing the user by its own name within that partial makes it much clearer that `title` is a separate object to the user itself.