Automatically apply application-level config (#176)
This includes a new handling of the “paths” setting for application views: if the application’s own “paths” setting is a relative path, then it will now be appended onto the application provider’s root path (e.g. the slice or application root paths).

This makes it possible to configure a standard “sub-path” at the application level that will then work for views in any given slice.