Create ApplicationContext and (initial) context helpers for use within full Hanami apps (#177)
Adds the beginning of seamless application integration to `Hanami::View::Context`.

Like we already do for `Hanami::View` and `Hanami::Action`, now when you inherit from `Hanami::View::Context` from within an Hanami application module, you'll get extra application-specific behaviours provided automatically.

To start with, this is reader methods for the `#request`, `#session`, and `#flash` (which will work with the `request` and `response` objects already provided as part of the `#view_context_options` in `Hanami::Action`'s view rendering integration.

No fancy metaprogramming required for this one, just a standard module with some instance methods.

n.b. this is certainly _not_ the final endpoint of the application context behavior, but it's a helpful starting point and required for application authors to fully access the features provided via the action/view rendering integration.

Along with this, also provide an experimental new `Hanami::View::ContextHelpers::ContentHelpers` module, which can be explicitly included into an application's context class to opt into some extra behavior on the application's view context. I imagine this pattern could be extended to provide an entire "standard library" of opt-in view context behavior. If this proves a useful approach, I'd imagine we could add some class-level niceties to make it easier to opt in without having to provide the fully qualified module names, but in the meantime, this is a simpler way to test the waters.