Use a reader effect to access the render env
This removes the cyclic dependency between a render environment and its context, part builder, and scope builder.

It also means we don’t need to create new copies of the context, part builder, and scope for each render environment we create (typically, there are 3 environments for a given rendering: a base environment, then a template- and layout-specific environment).

This is a particularly helpful in the case of the context object, because application authors may want to set state in the context object from within the template rendering, to make it available when the layout renders. This was not possible to do until now because both the template and layout renderings had different context objects.