Configure views with application inflector
Introduce Hanami::View::Application configuration, which will be used by the Hanami framework for Hanami.application.config.views.

This is a wrapper around the standalone view config (i.e. `Hanami::View.config`), with application-specific defaults applied, as well as configuration of the inflector disabled.

Along with this, update ApplicationView to configure itself with the Hanami application-level inflector, to ensure consistent inflection rules are used across all parts of the application.