Configure template on all application views
Previously, the `template` setting was only configured on _subclasses_ of an application view, presuming that the initial direct subclass of `Hanami::View` was an abstract “base” view, not to used directly.

This was an overly constricting approach, since there are numerous reasons that an application author may want to inherit directly from `Hanami::View` and use that subclass as a fully renderable view.

This change ensures the `template` setting is configured on _all_ descendants of `Hanami::View`.