Make PartBuilder and ScopeBuilder stateless
This greatly simplifies these objects and allows them to be set once and used everywhere, rather than being initialized for every view.

Passing the `rendering` into their call method still allows access to all the things they need, such as the view's config (which is now a public method on the rendering) so they can access the configured part_namespace and scope_namespace, as well as the inflector.