Simplify Context and streamline usage
Change how the context is prepared for rendering by duping the object in a new, adjusted `#dup_for_rendering` method.

This also makes it possible for users to fully own the `#initialize` method for their own context subclasses. They no longer need to call `super` inside their initialize, which is easy to forget and may be source of subtle bugs.