Finalize config in View#initialize
This ensures unexpected changes to config cannot be made after views have begun to be used.

In addition, finalizing the config will memoize its hash value, which makes it much faster to return. Update a couple of cache keys in Renderer to take advantage of this and improve rendering performance.