Include methods from helper modules as public
Replace `module_function` with `extend self` to allow helper methods to remain directly usable on their modules, but to also let those methods remain public when the module is included in a class.

This will allow an intermediary `helpers` object to be created for view parts (inside full Hanami apps) to access the standard helpers.