Include the used format in Rendered
This makes it possible for the caller of a view to inspect the resulting Rendered and determine what format was used for the rendering of the view. This is useful when the caller doesn’t have direct control over the view’s arguments, like when views are rendered in a Hanami action.