Remove LayoutNotFoundError and raise TemplateNotFoundError in all cases (#247)
Remove the specialized LayoutNotFoundError, because this was raising incorrect error if there were _missing partials_ being rendered inside the layout.

It is better that we raise a genuine, non-misleading error in these cases. The LayoutNotFoundError only provided marginal utility anyway, and the inclusion of the “layouts/“ prefix for a TemplateNotFoundError raised for a missing layout will still make it clear where the template was being looked for.